PLY files

+++ src/geometry/ray.rs
@@ -1,15 +1,15 @@
-use super::vec3::Vec3;
+use glam::Vec3;
 
 pub struct Ray {
     pub position: Vec3,
     pub direction: Vec3,
-    pub wavelength: f64,
-    pub strength: f64
+    pub wavelength: f32,
+    pub strength: f32
 }
 
 impl Ray {
 
-    pub fn new(position: Vec3, direction: Vec3, wavelength: f64, strength: f64) -> Ray {
+    pub fn new(position: Vec3, direction: Vec3, wavelength: f32, strength: f32) -> Ray {
         Ray { position, direction, wavelength, strength }
     }
 }
